+++ app/controllers/api/v1/accreditation_center/domains_controller.rb
@@ -4,15 +4,15 @@ module Api
   module V1
     module AccreditationCenter
       class DomainsController < BaseController
+        api :GET, 'api/v1/accreditation_center/domains/:name'
+        desc 'get domain by name'
         def show
           @domain = Domain.find_by(name: params[:name])
 
           if @domain
-            render json: {
-              domain: Serializers::Repp::Domain.new(@domain, sponsored: true).to_json
-            }, status: :ok
+            render_success(data: { domain: Serializers::Repp::Domain.new(@domain, sponsored: true).to_json })
           else
-            render json: { errors: 'Domain not found' }, status: :not_found
+            render_error('Domain not found', :not_found)
           end
         end
       end