+++ db/migrate/20260406125446_add_accred_status_to_registrars.rb
@@ -1,0 +1,9 @@
+class AddAccredStatusToRegistrars < ActiveRecord::Migration[6.1]
+  disable_ddl_transaction!
+
+  def change
+    add_column :registrars, :accreditation_date, :datetime unless column_exists?(:registrars, :accreditation_date)
+    add_column :registrars, :accreditation_expire_date, :datetime unless column_exists?(:registrars, :accreditation_expire_date)
+    add_index :registrars, :accreditation_date, algorithm: :concurrently unless index_exists?(:registrars, :accreditation_date)
+  end
+end